+++ src/Security/PropertyChecker.php
@@ -4,7 +4,7 @@ namespace App\Security;
 
 use App\Contract\NodeElementInterface;
 use App\Contract\RelationElementInterface;
-use App\Event\CheckUserSuppliedPropertiesEvent;
+// use App\Event\CheckUserSuppliedPropertiesEvent;
 use Psr\EventDispatcher\EventDispatcherInterface;
 
 class PropertyChecker
@@ -21,7 +21,7 @@ class PropertyChecker
      */
     public function runCheckUserSuppliedProperties(NodeElementInterface|RelationElementInterface $element, array $newProperties = []): void
     {
-        $event = new CheckUserSuppliedPropertiesEvent($element, $newProperties);
-        $this->eventDispatcher->dispatch($event);
+        //        $event = new CheckUserSuppliedPropertiesEvent($element, $newProperties);
+        //        $this->eventDispatcher->dispatch($event);
     }
 }