+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Switch datetime format from temporary string type to datetime.
 - Change default JSON encoding options.
 - Replaced configuration, documentation WIP.
+- Rework event system.
 
 ## 0.0.19 - 2023-08-10
 ### Added